openraft_macros

Macro expand

expand!() { /* proc-macro */ }

Render a template with arguments multiple times.

The template to expand is defined as (K,V) => { ... }, where K and V are template variables.

The template must contain at least 1 variable.

If the first macro argument is KEYED, the first variable serve as the key for deduplication. Otherwise, the first macro argument should be !KEYED, and no deduplication will be performed.

Example: KEYED for deduplication

The following code builds a series of let statements:

expand!(
    KEYED,
    // Template with variables K and V, and template body, excluding the braces.
    (K, T, V) => {let K: T = V;},
    // Arguments for rendering the template
    (a, u64, 1),
    (b, String, "foo".to_string()),
    (a, u32, 2), // duplicate a will be ignored
    (c, Vec<u8>, vec![1,2])
);

The above code will be transformed into:

let a: u64 = 1;
let b: String = "foo".to_string();
let c: Vec<u8> = vec![1, 2];

Example: !KEYED for no deduplication

expand!(!KEYED, (K, T, V) => {let K: T = V;},
                (c, u8, 8),
                (c, u16, 16),
);

The above code will be transformed into:

let c: u8 = 8;
let c: u16 = 16;